@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { FibonacciPage } from "../fibonacci-page/fibonacci-page";
 import { ListPage } from "../list-page/list-page";
 import { MainPage } from "../main-page/main-page";
@@ -13,29 +13,22 @@ function App() {
   return (
     <div className={appStyle.app}>
       <BrowserRouter>
-        <Switch>
-          <Route path="/" exact>
-            <MainPage />
+        <Routes>
+          <Route path="/" element={<MainPage />}>
           </Route>
-          <Route path="/recursion">
-            <StringComponent />
+          <Route path="/recursion" element={<StringComponent />}>
           </Route>
-          <Route path="/fibonacci">
-            <FibonacciPage />
+          <Route path="/fibonacci" element={<FibonacciPage />}>
           </Route>
-          <Route path="/sorting">
-            <SortingPage />
+          <Route path="/sorting" element={<SortingPage />}>
           </Route>
-          <Route path="/stack">
-            <StackPage />
+          <Route path="/stack" element={<StackPage />}>
           </Route>
-          <Route path="/queue">
-            <QueuePage />
+          <Route path="/queue" element={<QueuePage />}>
           </Route>
-          <Route path="/list">
-            <ListPage />
+          <Route path="/list" element={<ListPage />}>
           </Route>
-        </Switch>
+        </Routes>
       </BrowserRouter>
     </div>
   );
